requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.1
openai==1.14.0
python-dotenv==1.0.1
openpyxl==3.1.2